Moved pathfinding logic to a different file
